Struct alloy_consensus_any::AnyHeader

pub struct AnyHeader {

    pub parent_hash: B256,
    pub ommers_hash: B256,
    pub beneficiary: Address,
    pub state_root: B256,
    pub transactions_root: B256,
    pub receipts_root: B256,
    pub logs_bloom: Bloom,
    pub difficulty: U256,
    pub number: u64,
    pub gas_limit: u64,
    pub gas_used: u64,
    pub timestamp: u64,
    pub extra_data: Bytes,
    pub mix_hash: Option<B256>,
    pub nonce: Option<B64>,
    pub base_fee_per_gas: Option<u64>,
    pub withdrawals_root: Option<B256>,
    pub blob_gas_used: Option<u64>,
    pub excess_blob_gas: Option<u64>,
    pub parent_beacon_block_root: Option<B256>,
    pub requests_hash: Option<B256>,

}

Block header representation with certain fields made optional to account for possible differencies in network implementations.

Fields
parent_hash: B256

Hash of the parent

ommers_hash: B256

Hash of the uncles

beneficiary: Address

Alias of author

state_root: B256

State root hash

transactions_root: B256

Transactions root hash

receipts_root: B256

Transactions receipts root hash

logs_bloom: Bloom

Logs bloom

difficulty: U256

Difficulty

number: u64

Block number

gas_limit: u64

Gas Limit

gas_used: u64

Gas Used

timestamp: u64

Timestamp

extra_data: Bytes

Extra data

mix_hash: Option<B256>

Mix Hash

Before the merge this proves, combined with the nonce, that a sufficient amount of computation has been carried out on this block: the Proof-of-Work (PoF).

After the merge this is prevRandao: Randomness value for the generated payload.

This is an Option because it is not always set by non-ethereum networks.

See also https://eips.ethereum.org/EIPS/eip-4399 And https://github.com/ethereum/execution-apis/issues/328

nonce: Option<B64>

Nonce

base_fee_per_gas: Option<u64>

Base fee per unit of gas (if past London)

withdrawals_root: Option<B256>

Withdrawals root hash added by EIP-4895 and is ignored in legacy headers.

blob_gas_used: Option<u64>

Blob gas used

excess_blob_gas: Option<u64>

Excess blob gas

parent_beacon_block_root: Option<B256>

EIP-4788 parent beacon block root

requests_hash: Option<B256>

EIP-7685 requests hash.

Implementations

impl AnyHeader

pub fn try_into_header(self) -> Result<Header, Self>

Attempts to convert this header into a Header.

This can fail if the header is missing required fields:

nonce

mix_hash

If the conversion fails, the original AnyHeader is returned.

pub fn into_header_with_defaults(self) -> Header

Converts this header into a Header with default values for missing mandatory fields:

mix_hash

nonce